Fixes card-26 : middleware in separate file
Card url : https://trello.com/c/2N7pyTTQ
